persian_finance_manager/requirements.txt
python-telegram-bot[job-queue]
python-dotenv

# Web Panel Dependencies
fastapi
uvicorn[standard]
SQLAlchemy
Jinja2
python-multipart